preloading assets, utility funcs and local uri broken out
