--- examples/basic/Assets.js
@@ -1,11 +1,0 @@
-export default {
-  'icon.png': require(`./assets/icon.png`),
-  files: {
-    'bullet.png': require(`./assets/files/bullet.png`),
-    'enemy-bullet.png': require(`./assets/files/enemy-bullet.png`),
-    'explode.png': require(`./assets/files/explode.png`),
-    'invader32x32x4.png': require(`./assets/files/invader32x32x4.png`),
-    'player.png': require(`./assets/files/player.png`),
-    'starfield.png': require(`./assets/files/starfield.png`),
-  },
-};
