+++ examples/basic/utils/library.js
@@ -1,0 +1,5 @@
+import func from './functions';
+import images from './preloadImages';
+
+export { func };
+export { images };
