+++ examples/basic/utils/preloadImages.js
@@ -1,0 +1,22 @@
+const icon = require('../assets/icon.png');
+
+const bullet = require('../assets/files/bullet.png');
+const enemyBullet = require('../assets/files/enemy-bullet.png');
+const explode = require('../assets/files/explode.png');
+const invader = require('../assets/files/invader32x32x4.png');
+const player = require('../assets/files/player.png');
+const starfield = require('../assets/files/starfield.png');
+
+export default {
+  icon,
+
+  // game assets
+  files: {
+    bullet,
+    enemyBullet,
+    explode,
+    invader,
+    player,
+    starfield
+  }
+};
